backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
arq==0.26.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.9.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.1
reportlab==4.4.4
rsa==4.9.1
six==1.17.0
prometheus-client==0.16.0
sniffio==1.3.1
SQLAlchemy==2.0.47
sentry-sdk==1.19.1
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-redis==0.39b0
starlette==0.35.1
typing_extensions==4.15.0
uvicorn==0.27.0
watchfiles==1.1.1
websockets==16.0
pyotp==2.9.0
locust==2.32.2
edge-tts>=6.1.9
